2.1.1: Return time month to be integer

+++ src/io/regular.rs
@@ -8,7 +8,7 @@ fn get_date() -> impl core::fmt::Display {
     impl core::fmt::Display for TimeDate {
         #[inline(always)]
         fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
-            write!(f, "{}-{:02}-{:02} {:02}:{:02}:{:02}", self.0.year(), self.0.month(), self.0.day(), self.0.hour(), self.0.minute(), self.0.second())
+            write!(f, "{}-{:02}-{:02} {:02}:{:02}:{:02}", self.0.year(), self.0.month() as u8, self.0.day(), self.0.hour(), self.0.minute(), self.0.second())
         }
     }
